Update hacsintall.md for newer HACS versions
Made changes to reflect updates in Home Assistant and HACS since this guide was written. The main difference is that HACS will now automatically add the location of the js module to the resources, so this doesn't need to be done manually. 

I've left in the instructions of how to do this manually in a troubleshooting section, in case people are still using older versions. I have no idea when this change was made. 

There have also been some Home Assistant UI changes that are reflected in this section. Configuration is now called Settings, and Lovelace Dashboards is now simply Dashboards. Based on the initial documentation, I think Resources used to be a tab on this page, but it's now hidden behind a 3 dot menu in the top right. I've tried to cover all bases by leaving the initial instructions as written, while also including instructions for the updated UI.